+++ src/test/java/test/EmployeeSearchCondition.java
@@ -1,0 +1,27 @@
+package test;
+
+public class EmployeeSearchCondition {
+
+	private Integer salaryMin;
+
+	private Integer salaryMax;
+
+	public EmployeeSearchCondition() {
+	}
+
+	public Integer getSalaryMin() {
+		return salaryMin;
+	}
+
+	public void setSalaryMin(Integer salaryMin) {
+		this.salaryMin = salaryMin;
+	}
+
+	public Integer getSalaryMax() {
+		return salaryMax;
+	}
+
+	public void setSalaryMax(Integer salaryMax) {
+		this.salaryMax = salaryMax;
+	}
+}